cargo_auto_local_lib
Module file_hashes_mod

Calculate file hashes.

File hashes are used to check if some file has changed. Then we can run commands like compile only if a file has changed.

Structs

FileHashes

The struct represents the file automation_tasks_rs/._file_hashes.json.

FileMetaData

Struct with file metadata.

Functions

are_all_files_equal 🔒

Check if all files are equal.

is_project_changed

Check if the files are modified in automation_tasks_rs.

read_file_metadata

Make a vector of file metadata.

read_json_file 🔒

Read automation_tasks_rs/.file_hashes.json.

save_json_file_for_file_meta_data

Save the new file metadata.

sha256_digest 🔒

Calculate the hash for a file.