Module cargo_auto_local_lib::file_hashes_mod

calculate file hashes

File hashes are used to check if some file has changed. Then we can run commands like compile only if a file has changed.

Structs

FileHashes

the struct that represents the file automation_tasks_rs/._file_hashes.json

FileMetaData

file metadata

Functions

are_all_files_equal 🔒

is_project_changed

check if the files are modified and in automation_tasks_rs The modified date of files is not usable when using git. The checkout will make dates newer than they really are. I should use a hash of files and write them in the same directory for later comparison.

read_file_metadata

make a vector of file metadata

read_json_file 🔒

read automation_tasks_rs/.file_hashes.json

save_json_file_for_file_meta_data

save the new file metadata

sha256_digest 🔒

calculate the hash for a file